requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# JWT and crypto
PyJWT>=2.8.0
cryptography>=42.0.0

# HTTP client
httpx>=0.26.0

# CORS
python-jose[cryptography]>=3.3.0
